First stage of echoServer work, fix bugs in gossipServer
echoServer is currently disabled - needs more dev work but getting there
This will enable the gossipServer peers to be shared and stateful. Also
allows for testing of multicast per node.

Added a bunch of extra tests in the integration tests. Now checks that
all shared state between nodes is the same, and that there are no
double up of ports or missing ports etc.

Fixed a reasonable bug where state was being overwritten if it was
received inbetween startup and the discover reply. This has helped
with far fewer NAKs (maybe none) being received.

Moved to enums for messages, means less data transfer for messages. Will
look at moving some other types to enums as well.

There is still a small issue where the odd node is wiping state before it
is being sent to the observer. This is only present in observer mode,
so no impact, will fix in next release. In the last test run, this
effected 1/50 runs.

Began work on echoServer tests.